refactor: KanbanModal 을 KanbanTodo 로 부터 분리 [#86]
