+++ client/src/components/KanbanColumn/index.tsx
@@ -1,9 +1,7 @@
 import React, { useState, useRef } from 'react';
 import Styled from '@/components/KanbanColumn/style';
 import { StoryType } from '@/types/story';
-import { deleteStoryWitId } from '@/lib/api/story';
 import { useStoryState, useStoryDispatch } from '@/lib/hooks/useContextHooks';
-import { Modal } from '@/lib/design';
 import { KanbanItem, KanbanAddBtn } from '@/components';
 
 interface KanbanProps {
@@ -15,15 +13,8 @@ type CopyList = [StoryType];
 const KanbanColumn = ({ projectId }: KanbanProps) => {
   const storyList = useStoryState();
   const dispatchStory = useStoryDispatch();
-  const [showModal, setShowModal] = useState(false);
-  const [shouldDeleteKey, setDeleteKey] = useState(0);
   const draggingItem = useRef<number | null>(0);
   const dragoverItem = useRef<number | null>(0);
-
-  const deleteStory = async () => {
-    dispatchStory({ type: 'REMOVE_STORY', id: shouldDeleteKey });
-    await deleteStoryWitId(shouldDeleteKey);
-  };
 
   const handleDragStart = (e: React.SyntheticEvent<HTMLElement>, position: number) => {
     draggingItem.current = position;
@@ -51,21 +42,11 @@ const KanbanColumn = ({ projectId }: KanbanProps) => {
           key={story.id}
           index={index}
           story={story}
-          setDeleteKey={setDeleteKey}
-          setShowModal={setShowModal}
           handleDragStart={handleDragStart}
           handleDragEnter={handleDragEnter}
         />
       ))}
       <KanbanAddBtn projectId={projectId} />
-      <Modal
-        shouldConfirm
-        title={'스토리를 삭제하시겠습니까?'}
-        visible={showModal}
-        onClose={() => setShowModal(false)}
-        onClickCancel={() => setShowModal(false)}
-        onClickOk={deleteStory}
-      />
     </Styled.Column>
   );
 };
