+++ client/src/components/KanbanModal/index.tsx
@@ -1,0 +1,38 @@
+import React, { useState, createContext } from 'react';
+import { Modal } from '@/lib/design';
+import { deleteStoryWitId } from '@/lib/api/story';
+import { useStoryDispatch } from '@/lib/hooks/useContextHooks';
+
+export type ModalContextType = {
+  setShowModal: (args: boolean) => void;
+  setDeleteItem: (arg: number) => void;
+};
+
+export const KanbanModalContext = createContext<ModalContextType | null>(null);
+
+const KanbanModal = ({ children }: { children: React.ReactNode }) => {
+  const [showModal, setShowModal] = useState(false);
+  const [shouldDeleteKey, setDeleteItem] = useState(0);
+
+  const dispatchStory = useStoryDispatch();
+  const deleteStory = async () => {
+    dispatchStory({ type: 'REMOVE_STORY', id: shouldDeleteKey });
+    await deleteStoryWitId(shouldDeleteKey);
+  };
+
+  return (
+    <KanbanModalContext.Provider value={{ setShowModal, setDeleteItem }}>
+      <Modal
+        shouldConfirm
+        title={'스토리를 삭제하시겠습니까?'}
+        visible={showModal}
+        onClose={() => setShowModal(false)}
+        onClickCancel={() => setShowModal(false)}
+        onClickOk={deleteStory}
+      />
+      {children}
+    </KanbanModalContext.Provider>
+  );
+};
+
+export default KanbanModal;
